requirements.txt
agentfield>=0.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
aiohttp>=3.9.0
